requirements.txt
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cld2-cffi==0.1.4
feedparser==5.2.1
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
httplib2==0.17.0
idna==2.9
multi-rake==0.0.1
nltk==3.4.5
numpy==1.18.1
Pillow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pyrsistent==0.15.7
regex==2020.2.20
requests==2.23.0
rsa==4.0
six==1.14.0
uritemplate==3.0.1
urllib3==1.25.8
Wikipedia-API==0.5.4
